sharedPref with different users working

+++ app/src/main/java/com/kodbale/dkode/Database/Question.java
@@ -72,7 +72,7 @@ public class Question {
         return mIsText;
     }
 
-    public long getQuestionId() {
+    public int getQuestionId() {
         return mQuestionId;
     }
